Pull the Calm item transformer into a separate trait
